Development/add http server (#4)
* HTTP server added and working

It implements login by generating and sending token to client
It is a standard practice and this token is saved in browser history
On subsequent request that tooken is provided by user
This way user can also not directly bypass and open dashboard and ota
because server checks the token header, if not, it redirects to login

Right now the session time is not used. if implemented the token will
expire after some time

* Fix/ota redirecting fix (#2)

* about to add 'bearer' to all token headers

* Page working with authenticated access

If user access a page directly without login, 401 UnAuth is returned
and the js of that page redirects to the login page
On login an Auth token is generated which is sent as json body
and js saves it to browser data. on subsequent request that token is
retreived and passed on evenry request
The upload  is also working
The timer of the token is not used right now but must be added in future

* Working verfied. Provides the pages after checking logged in status

If pages like ota and dashboard are directly accessed without logged in
It redirects to the login pages
There is alot of redundancy which must be optimized. for example the
redirect to login is no used, because the js is programmed to redirect
itself onn 401 unauth response

* Sub branching approach to be used now onwards

* event-adapter added

* Bugfix: The sess duration was set to 1800 ms, changed to 180000ms

That is why the subsequent pages were abrublty redirecting to login page
because the session expired in just 1.8 seconds

* Checks if already logged in and redirects to the dashboard page

The checkauth logic from the oher pages added and action on document loading added

On load it checks if there is already any token and then sends to server for auth

if success then redirects to dashboard page

* Bugfix: The logout call missed the "Bearer" prefix to toke
Now added it

* Fix/array pool return fix (#3)

* Sends file as a single post instead of multiple

Earlier a simulated chunk transfer scheme was used which is wrong
There is no mainstream standard chunk transfer from client to server

Chunk transfer is used when data size is not  known upfront
Now the file is sent as a single POST request

* A new data type added whose instance is sent in the event post

That data type is used by the routine event handler to get its memebers

Its members are buffer pointer and the length, and it is meant for ota

So current data received

* File upload handler changed and the get buffer method signature changed

The get buffer now supports portMaxDelay if the wait_time argument has 0 value

So now it waits for portMaxDelay for receving buffer
Also the data posted is a struct whose one member is the buffer pointer and
the other member is length